Changed to using std::thread, and added some comments on used libs.
The only gotcha here is that std::thread doesn't implicitly do
thread::detach (e.g. when it goes out of scope and gets RAIId),
which was causing the bug that made me stick with boost::thread
in the first place. Fixed this. Also added some TODOs around
thread safety (file name collision is still possible, but
it's not a big deal since it's triggered by user input).